added incr/2, decr/2, addcounter/1, literal/1 incr/2 implements the increment command for memcached, decr implements the decrement command.
addcounter/1 uses a simple (non-CAS) set to initialized a counter set
at zero, which can then be fiddled with via incr/2 and decr/2. This
was necessary because merle functions mainly as a key/value store for
arbitrary erlang data instead of a general data-store, such as we want
for cacti purposes.

literal/1 allows the user to submit literal strings over the socket to
memcached. This should be used with extreme caution as memcached is
intolerant of malformed input, but it is useful for determining what
strings sent to memcached should look like and checking assumptions
about its behavior.

TODO: (optional) need a Check-And-Set version of addcounter so that
distributed applications can update the same value with incr/2 and
decr/2. This would save client code from having to try changing a
value and then making a separate call to initialize it if it does not
exist.